Added matrix transpose function and testing
When creating shares from a large piece of data, a matrix is returned
with each row containing the shares to one byte of data. When handing
out shares of the data, you want each full share to contain one piece to
each byte of data. Transposing the returned matrix makes this simple,
with each row having a share for each byte of the secret.