Moved transpose operation into share functions
The transposition was originally done outside this implementation but
considering that this operation is necessary for almost all use cases,
it makes sense to move the transposition into the implementation. Now
when creating shares from a large secret, it returns a transposed list
of shares that are already ready for distribution and the reconstruction
implementation now expects the shares to come in in that same format.