Made the geometry module public
Since some of the functions in the share module require the Point
structure as an argument or return value, there were two choices:
 - Change the functions to accept BigInt and re-map them to Points
 within the function, and just return BigInt.
 - Make public the geometry module
I want clients to have as much control over the use of the library as
possible and have it be as flexible as possible. Forcing certain
operations to take place when they are not 100% necessary puts
constraints on how the library is used.