Renamed share to raw_share
The functional nature of this library prior to this point made it a bit
messy and not really fit in with Rust's object-oriented design. A
new wrapper around these functions will be added, but these functions
will still be public and useable if a user wants to use them directly or
create their own wrapper.