Fixed a continued issue with modulo op with negs
The modulo operation was still not outputting the right values when Self
was a negative value. The natural_mod_bigint function was removed and no
longer necessary and the main modulo function takes care of cases of
negative numbers as well properly. Also removed old unusued code that
was commented out.